Trabalho Folha pagamento - Interface init

+++ brasileirao2018Elizelton/src/main/java/utility/Dados.java
@@ -29,7 +29,7 @@ public class Dados {
                 analisa(jgLinha);
             }
             // Ordena
-            byte i = 0;
+            Integer i = 0;
             for (Time t : lstTimes) {
                 t.setClas(++i);
             }
